+++ Devsource/components/filterB.js
@@ -9,7 +9,7 @@ import {
 } from "react-native";
 import estilo from "../estilo";
 
-export default function FilterB({ ola, customBool, customCond, curVal }) {
+export default function FilterB({ ola, customBool, customCond, curVal, func }) {
   const [activo, setActivo] = useState(false);
 
   function geralDao() {
@@ -18,8 +18,15 @@ export default function FilterB({ ola, customBool, customCond, curVal }) {
   //geralDao();
   function actualizar(vdd) {
     //console.log("antes:" + vdd);
+
+    if (func != null) {
+      func();
+    }
+
+
     if (customBool == false) {
       setActivo(vdd);
+      console.log("ativo");
     }
     else {
       if (customCond == curVal) {
@@ -32,19 +39,24 @@ export default function FilterB({ ola, customBool, customCond, curVal }) {
     // console.log("Dps:" + vdd);
   }
 
+
+
+
   return (
     <Pressable
       onPress={() => {
         actualizar(!activo);
-        console.log("mist");
+        //console.log("mist");
       }}
     >
       <Text
         style={[
           estilo.filterButton,
-          activo
-            ? { backgroundColor: "darkorange", color: "white" }
-            : { backgroundColor: "#FFFFFF" },
+          customBool ? customCond == curVal ?
+            { backgroundColor: "darkorange", color: "white" }
+            : { backgroundColor: "#FFFFFF" } :
+            activo ? { backgroundColor: "darkorange", color: "white" }
+              : { backgroundColor: "#FFFFFF" },
         ]}
       >
         {ola}
